simplify inputs implementation
